+++ electron-vite-project/src/components/ScenesComponent.tsx
@@ -14,8 +14,6 @@
  */
 import { useState, useEffect, useContext } from 'react';
 import './ScenesComponent.css';
-import '../assets/GridLines';
-import GridLines from '../assets/GridLines';
 import { TranslationContext } from './TranslationContext';
 import { useConnectionContext } from '../components/ConnectionContext';
 
@@ -218,7 +216,7 @@ const ScenesComponent: React.FC<ScenesComponentProps> = ({ sideId, setAddScene, 
           style={{ height: `${height}px` }}
           onClick={handleAddScene}
         >
-          <GridLines height={height} />
+          <div className={`sceneBorder ${sideId === 1 ? 'borderBig' : 'borderSmall'}`}></div>
           {!buttonDisabled && <div className='AddSceneIcon'></div>}
           <div className='AddSceneButtonFont'>{buttonText}</div>
         </button>
@@ -228,7 +226,7 @@ const ScenesComponent: React.FC<ScenesComponentProps> = ({ sideId, setAddScene, 
             style={{ height: `${height}px` }}
             key={scene.id}
           >
-            <GridLines height={height} />
+            <div className={`sceneBorder ${sideId === 1 ? 'borderBig' : 'borderSmall'}`}></div>
           </div>
         ))}
       </div>